Extract parseCommandLine from main.js to its own file for testing
